en_itn/requirements_nlp.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==20.3.0
boto3==1.17.55
botocore==1.20.55
braceexpand==0.1.6
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
filelock==3.0.12
fsspec==2021.4.0
future==0.18.2
gdown==3.12.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
h5py==3.2.1
hydra-core==1.0.6
idna==2.10
importlib-resources==5.1.2
inflect==5.3.0
ipadic==1.0.0
jieba==0.42.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.1
mecab-python3==1.0.3
megatron-lm==2.2.0
multidict==5.1.0
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@fbc5c2b4033896b75f643cf169976c1d38505d65
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.0
omegaconf==2.0.6
onnx==1.9.0
OpenCC==1.1.2
packaging==20.9
pangu==4.0.6.1
Pillow==8.2.0
portalocker==2.0.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pynini @ file:///home/conda/feedstock_root/build_artifacts/pynini_1605833195807/work
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.2.8
PyYAML==5.3.1
rapidfuzz==1.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.1
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.2
sentencepiece==0.1.95
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
torchmetrics==0.3.1
tqdm==4.60.0
transformers==4.5.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.4
webdataset==0.1.59
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
yarl==1.6.3
youtokentome==1.0.6
streamlit==0.80.0